+++ x/pki/types/message_update_pki_revocation_distribution_point.go
@@ -9,7 +9,7 @@ const TypeMsgUpdatePkiRevocationDistributionPoint = "update_pki_revocation_distr
 
 var _ sdk.Msg = &MsgUpdatePkiRevocationDistributionPoint{}
 
-func NewMsgUpdatePkiRevocationDistributionPoint(signer string, vid uint64, label string, crlSignerCertificate string, issuerSubjectKeyID string, dataUrl string, dataFileSize uint64, dataDigest string, dataDigestType uint64) *MsgUpdatePkiRevocationDistributionPoint {
+func NewMsgUpdatePkiRevocationDistributionPoint(signer string, vid int32, label string, crlSignerCertificate string, issuerSubjectKeyID string, dataUrl string, dataFileSize uint64, dataDigest string, dataDigestType uint32) *MsgUpdatePkiRevocationDistributionPoint {
 	return &MsgUpdatePkiRevocationDistributionPoint{
 		Signer:               signer,
 		Vid:                  vid,
